+++ admin_test.go
@@ -1,6 +1,7 @@
 package gadm
 
 import (
+	"io"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -29,4 +30,13 @@ func TestText(t *testing.T) {
 	gotext.Configure("translations", "zh_Hant_TW", "admin")
 	is.Equal("首頁", gotext.Get("Home"))
 	is.Equal(`檔案 "foo" 已經存在。`, gotext.Get(`File "%s" already exists.`, "foo"))
+}
+
+func TestGen(t *testing.T) {
+	is := assert.New(t)
+
+	g := NewGenerator("sqlite:examples/sqla/sample.db")
+
+	err := g.Run(nil, io.Discard)
+	is.Nil(err)
 }
