--- templates/actions.tmpl
@@ -1,31 +1,0 @@
-{{/* import 'admin/static.html' as admin_static with context */}}
-
-{{/* (actions, btn_class='nav-link dropdown-toggle' */}}
-{{ define "actionlib_dropdown" -}}
-  {{ $btn_class := "nav-link dropdown-toggle" }}
-    <a class="{{ $btn_class }}" data-toggle="dropdown" href="javascript:void(0)" role="button" aria-haspopup="true"
-       aria-expanded="false">{{ gettext "With selected" }}<b class="caret"></b></a>
-    <div class="dropdown-menu">
-            <a class="dropdown-item" href="javascript:void(0)"
-               onclick="return modelActions.execute('{{ .Name }}');">{{ .Title }}</a>
-    </div>
-{{ end }}
-
-{{/* (actions, url) */}}
-{{ define "actionlib_form" }}
-        <form id="action_form" action="{{ .URL }}" method="POST" style="display: none">
-        {{ if .CSRFToken }}
-            <input type="hidden" name="csrf_token" value="{{ .CSRFToken }}"/>
-        {{ end }}
-            {{/* action_form.url(value=return_url) */}}
-            <input id="url" name="url" type="hidden" value="{{ .ReturnURL }}">
-            <input id="action" name="action" type="hidden" value="{{ .Name }}">
-        </form>
-{{ end }}
-
-
-{{ define "actionlib_script" }}
-        <div id="actions-confirmation-data" style="display:none;">{{ marshal .actions_confirmation }}</div>
-        <div id="message-data" style="display:none;">{{ json .message  }}</div>
-        <script src="{{ admin_static_url "admin/js/actions.js" "1.0.0" }}"></script>
-{{ end }}
