+++ templates/lib.tmpl
@@ -150,6 +150,7 @@
     {{ end }}
 {{ end }}
 
+
 {{ define "render_form" }}
   <form action="{{ .form.action }}" method="POST" role="form" class="admin-form" enctype="multipart/form-data">
     <fieldset>
@@ -165,8 +166,6 @@
       | args}}
   </form>
 {{ end }}
-
-
 
 
 {{ define "form_css" }}
@@ -222,17 +221,3 @@
   {{- end -}}
 {{ end }}
 
-
-{{ define "admin_nav" }}
-  <ul class="nav nav-tabs">
-      <li class="nav-item">
-          <a href="{{get_url ".console"}}" class="nav-link {{if eq .path (get_url ".console") }}active{{end}}">SQL Console</a>
-      </li>
-      <li class="nav-item">
-          <a href="{{get_url ".trace"}}" title="Trace" class="nav-link {{if eq .path (get_url ".trace") }}active{{end}}">Trace</a>
-      </li>
-      <li class="nav-item">
-          <a href="{{get_url ".generate"}}" title="Generate from Database" class="nav-link {{if eq .path (get_url ".generate") }}active{{end}}">Generate</a>
-      </li>
-  </ul>
-{{ end }}
