+++ templates/model_create.tmpl
@@ -1,5 +1,4 @@
 {{ template "master.tmpl" .}}
-{{ template "actions.tmpl" .}}
 {{ template "lib.tmpl" .}}
